services/reviewer/requirements.txt
# ------------ Pydantic & Settings ------------
pydantic==2.11.7
pydantic-settings==2.10.1

# ------------ Broker Messaging ------------
aiokafka==0.12.0

# ------------ Utilities & Other ------------
colorlog==6.9.0
dependency_injector==4.48.1
httpx==0.28.1
gitpython==3.1.45
tenacity==9.1.2
tiktoken==0.11.0
